# api/openapi.yaml
#
#  ========================LICENSE_START=================================
#   Copyright (C) 2024-2025: Deutsche Telekom
#
#   Licensed under the Apache License, Version 2.0 (the "License");
#   you may not use this file except in compliance with the License.
#   You may obtain a copy of the License at
#
#        http://www.apache.org/licenses/LICENSE-2.0
#
#   Unless required by applicable law or agreed to in writing, software
#   distributed under the License is distributed on an "AS IS" BASIS,
#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#   See the License for the specific language governing permissions and
#  limitations under the License.
#  SPDX-License-Identifier: Apache-2.0
#  ========================LICENSE_END===================================
#
openapi: 3.0.3
info:
  title: "Policy OPA PDP Documentation"
  description: Policy OPA PDP Service
  version: 1.0.3
  x-component: Policy Framework
  x-planned-retirement-date: tbd
  contact:
      name: Deena Mukundan
      email: dm00536893@techmahindra.com
servers:
- url: http://policy-opa-pdp:8282/policy/pdpo/v1
- url: https://policy-opa-pdp:8282/policy/pdpo/v1
tags:
- name: Decision
- name: Statistics
- name: HealthCheck
- name: Dynamic Data Update
paths:
  /decision:
    post:
      tags:
      - Decision
      summary: Fetch the decision using specified decision parameters
      description: Returns the policy decision from Policy OPA PDP
      operationId: decision
      parameters:
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OPADecisionRequest'
          application/yaml:
            schema:
              $ref: '#/components/schemas/OPADecisionRequest'
        required: true
      responses:
        200:
          description: successful operation
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              schema:
                type: string
            X-PatchVersion:
              description: Used only to communicate a PATCH version in a response
                for troubleshooting purposes only, and will not be provided by the
                client on request
              schema:
                type: string
            X-MinorVersion:
              description: Used to request or communicate a MINOR version back from
                the client to the server, and from the server back to the client
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPADecisionResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/OPADecisionResponse'

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Authentication Error
          content: {}
        403:
          description: Authorization Error
          content: {}
        500:
          description: Internal Server Error
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        last-mod-release: Paris
        pdpo-version: 1.0.3
      x-codegen-request-body-name: body
  /healthcheck:
    get:
      tags:
      - HealthCheck
      summary: Perform a system healthcheck
      description: Provides healthy status of the Policy OPA PDP component
      operationId: healthcheck
      parameters:
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: successful operation
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              schema:
                type: string
            X-PatchVersion:
              description: Used only to communicate a PATCH version in a response
                for troubleshooting purposes only, and will not be provided by the
                client on request
              schema:
                type: string
            X-MinorVersion:
              description: Used to request or communicate a MINOR version back from
                the client to the server, and from the server back to the client
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckReport'
            application/yaml:
              schema:
                $ref: '#/components/schemas/HealthCheckReport'
        401:
          description: Authentication Error
          content: {}
        403:
          description: Authorization Error
          content: {}
        500:
          description: Internal Server Error
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        last-mod-release: Paris
        pdpo-version: 1.0.3
  /statistics:
    get:
      tags:
      - Statistics
      summary: Fetch current statistics
      description: Provides current statistics of the Policy OPA PDP component 
      operationId: statistics
      parameters:
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: successful operation
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              schema:
                type: string
            X-PatchVersion:
              description: Used only to communicate a PATCH version in a response
                for troubleshooting purposes only, and will not be provided by the
                client on request
              schema:
                type: string
            X-MinorVersion:
              description: Used to request or communicate a MINOR version back from
                the client to the server, and from the server back to the client
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsReport'
            application/yaml:
              schema:
                $ref: '#/components/schemas/StatisticsReport'
        401:
          description: Authentication Error
          content: {}
        403:
          description: Authorization Error
          content: {}
        500:
          description: Internal Server Error
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        last-mod-release: Paris
        pdpo-version: 1.0.3
  /data/{path}:
    patch:
      tags:
      - Dynamic Data Update
      summary: Updates the data which is already added by a PUT request
      description: Returns the success or failure response
      operationId: patchdata
      parameters:
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      - name: path
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OPADataUpdateRequest'
          application/yaml:
            schema:
              $ref: '#/components/schemas/OPADataUpdateRequest'
        required: false
      responses:
        200:
          description: successful operation
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              schema:
                type: string
            X-PatchVersion:
              description: Used only to communicate a PATCH version in a response
                for troubleshooting purposes only, and will not be provided by the
                client on request
              schema:
                type: string
            X-MinorVersion:
              description: Used to request or communicate a MINOR version back from
                the client to the server, and from the server back to the client
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              schema:
                type: string
                format: uuid
          content: {}

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Authentication Error
          content: {}
        403:
          description: Authorization Error
          content: {}
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basicAuth: []
      x-interface info:
        last-mod-release: Paris
        pdpo-version: 1.0.3
      x-codegen-request-body-name: body
    get:
      tags:
      - Dynamic Data Update
      summary: Gets the data dynamically
      description: Returns success or failure
      operationId: data-get
      parameters:
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      - name: path
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: successful operation
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              schema:
                type: string
            X-PatchVersion:
              description: Used only to communicate a PATCH version in a response
                for troubleshooting purposes only, and will not be provided by the
                client on request
              schema:
                type: string
            X-MinorVersion:
              description: Used to request or communicate a MINOR version back from
                the client to the server, and from the server back to the client
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPADataResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/OPADataResponse'
        401:
          description: Authentication Error
          content: {}
        403:
          description: Authorization Error
          content: {}
        500:
          description: Internal Server Error
          content: {}
      security:
        - basicAuth: []
      x-interface info:
        last-mod-release: Paris
        pdpo-version: 1.0.3
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        responseCode:
          type: string
          enum:
          - internal_error
          - bad_request
          - evaluation_error
          - unauthorized
          - invalid_parameter
          - invalid_operation
          - resource_not_found
          - resource_conflict
          - undefined_document
          - opa_undefined_error
        errorMessage:
          type: string
        policyName:
          type: string
    OPADecisionRequest:
      type: object
      properties:
        onapName:
          type: string
        onapComponent:
          type: string
        onapInstance:
          type: string
        currentDateTime:
          type: string
          format: date-time
        currentDate:
          type: string
          format: date
        currentTime:
          type: string
          format: time
        timeZone:
          type: string
          description: "Timezone in IANA format (e.g., 'America/NewYork', 'Europe/Paris', 'UTC')"
        timeOffset:
          type: string
          pattern: '^[+-]?\d{2}:\d{2}$'
          description: "Time offset in hours and minutes, e.g., '+02:00' or '-05:00'"
        policyName:
          type: string
        policyFilter:
          type: array
          items:
            type: string
        input:
          anyOf:
           - x-go-type: "interface{}"
           - type: object
             additionalProperties: true
             example:
                    user: alice
                    action: read
                    object: id123
                    type: dog          
      required:
      - policyName
      - policyFilter
      - input
    OPADataUpdateRequest:
      type: object
      properties:
        onapName:
          type: string
        onapComponent:
          type: string
        onapInstance:
          type: string
        currentDateTime:
          type: string
          format: date-time
        currentDate:
          type: string
          format: date
        currentTime:
          type: string
          format: time
        timeZone:
          type: string
          description: "Timezone in IANA format (e.g., 'America/NewYork', 'Europe/Paris', 'UTC')"
        timeOffset:
          type: string
          pattern: '^[+-]?\d{2}:\d{2}$'
          description: "Time offset in hours and minutes, e.g., '+02:00' or '-05:00'"
        policyName:
          type: string
        data:
          type: array
          items:
             type: object
          example:
                -
                    user: alice
                    action: read
                    object: id123
                    type: dog
    HealthCheckReport:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        healthy:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
    OPADecisionResponse:
      type: object
      properties:
        policyName:
          type: string
        statusMessage:
          type: string
        output:
          type: object
          additionalProperties: true
      required:
      - policyName
      - output

    StatisticsReport:
      type: object
      properties:
        code:
          type: integer
          format: int32
        totalPolicyTypesCount:
          type: integer
          format: int64
        totalPoliciesCount:
          type: integer
          format: int64
        totalErrorCount:
          type: integer
          format: int64
        deploySuccessCount:
          type: integer
          format: int64
        deployFailureCount:
          type: integer
          format: int64
        undeploySuccessCount:
          type: integer
          format: int64
        undeployFailureCount:
          type: integer
          format: int64
        decisionSuccessCount:
          type: integer
          format: int64
        decisionFailureCount:
          type: integer
          format: int64
        dynamicDataUpdateSuccessCount:
          type: integer
          format: int64
        dynamicDataUpdateFailureCount:
          type: integer
          format: int64
    OPADataResponse:
      type: object
      properties:
         data:
           anyOf:
           - x-go-type: "interface{}"
           - type: object
             additionalProperties: true
  securitySchemes:
    basicAuth:
      type: http
      description: ""
      scheme: basic